refactor: melhora organização dos componentes
