+++ db.json
@@ -23,6 +23,22 @@
       "employeeName": "José Leão",
       "status": "REPROVED",
       "cpf": "78502270001"
+    },
+    {
+      "id": "16fa",
+      "admissionDate": "2002-02-27",
+      "email": "ltoningalvan@gmail.com",
+      "employeeName": "Luan Galvan",
+      "cpf": "045.670.040-43",
+      "status": "REVIEW"
+    },
+    {
+      "id": "5d39",
+      "admissionDate": "2002-02-27",
+      "email": "ltoningalvan@gmail.com",
+      "employeeName": "Luan Galvan",
+      "cpf": "045.670.040-43",
+      "status": "REVIEW"
     }
   ]
 }
