+++ src/components/Button/Button.types.ts
@@ -1,0 +1,5 @@
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  children: React.ReactNode;
+  isLoading?: boolean;
+}
