--- src/components/Buttons/index.test.tsx
@@ -1,10 +1,0 @@
-import Button from ".";
-import { render, screen } from "@testing-library/react";
-
-describe("Button", () => {
-  it("Should show button", () => {
-    const { debug } = render(<Button>Ativar</Button>);
-    expect(screen.getByRole("button", { name: /ativar/i }));
-    debug();
-  });
-});
