+++ src/components/IconButton/IconButton.styles.ts
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+export const IconButtonContainer = styled.button`
+  cursor: pointer;
+  border: 2px solid #64a98c;
+  width: fit-content;
+  padding: 4px;
+  border-radius: 24px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: transparent;
+  svg {
+    color: #64a98c;
+  }
+`;
