+++ src/components/IconButton/IconButton.types.ts
@@ -1,0 +1,3 @@
+export type IconButtonProps = {
+  children?: React.ReactNode;
+} & React.HTMLAttributes<HTMLButtonElement>;
