+++ src/components/TextField/TextField.types.ts
@@ -1,0 +1,6 @@
+import { InputHTMLAttributes } from "react";
+
+export type TextFieldProps = {
+  label?: string;
+  error?: string;
+} & InputHTMLAttributes<HTMLInputElement>;
