+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { default as Button } from "./Button/Button.view";
+export { default as Header } from "./Header/Header.view";
+export { default as IconButton } from "./IconButton/IconButton.view";
+export { default as TextField } from "./TextField/TextField.view";
