+++ src/pages/Dashboard/components/RegistrationCard/index.tsx
@@ -1,4 +1,4 @@
-import { ButtonSmall } from "~/components/Buttons";
+import styled from "styled-components";
 import * as S from "./styles";
 import {
   HiOutlineMail,
@@ -10,6 +10,20 @@ import {
 type Props = {
   data: any;
 };
+
+export const ButtonSmall = styled.button<{
+  bgcolor?: string;
+  color?: string;
+}>`
+  font-size: 12px;
+  outline: none;
+  border-radius: 4px;
+  border: none;
+  padding: 4px 16px;
+  background-color: ${(props) => props.bgcolor ?? "none"};
+  color: ${(props) => props.color ?? "#000"};
+  cursor: pointer;
+`;
 
 const RegistrationCard = (props: Props) => {
   return (
@@ -27,7 +41,7 @@ const RegistrationCard = (props: Props) => {
         <span>{props.data.admissionDate}</span>
       </S.IconAndText>
       <S.Actions>
-        <ButtonSmall bgcolor="rgb(255, 145, 154)" >Reprovar</ButtonSmall>
+        <ButtonSmall bgcolor="rgb(255, 145, 154)">Reprovar</ButtonSmall>
         <ButtonSmall bgcolor="rgb(155, 229, 155)">Aprovar</ButtonSmall>
         <ButtonSmall bgcolor="#ff8858">Revisar novamente</ButtonSmall>
