+++ src/pages/Dashboard/components/Searchbar/index.tsx
@@ -1,20 +1,20 @@
 import { HiRefresh } from "react-icons/hi";
 import { useHistory } from "react-router-dom";
-import Button from "~/components/Buttons";
-import { IconButton } from "~/components/Buttons/IconButton";
-import TextField from "~/components/TextField";
+import { Button, IconButton, TextField } from "~/components";
+
 import routes from "~/router/routes";
 import * as S from "./styles";
+
 export const SearchBar = () => {
   const history = useHistory();
 
   const goToNewAdmissionPage = () => {
     history.push(routes.newUser);
   };
-  
+
   return (
     <S.Container>
-      <TextField  placeholder="Digite um CPF válido" />
+      <TextField placeholder="Digite um CPF válido" />
       <S.Actions>
         <IconButton aria-label="refetch">
           <HiRefresh />
