+++ src/pages/NewUser/index.ts
