+++ src/router/index.tsx
@@ -1,7 +1,8 @@
 import { HashRouter, Redirect, Route, Switch } from "react-router-dom";
-import routes from "./routes";
-import DashboardPage from "~/pages/Dashboard";
+import { DashboardPage } from "~/pages/Dashboard";
 import { NewUserPage } from "~/pages/NewUser";
+
+import routes from "./routes";
 
 const Router = () => {
   return (
@@ -9,11 +10,6 @@ const Router = () => {
       <Switch>
         <Route exact path={routes.dashboard} component={DashboardPage} />
         <Route exact path={routes.newUser} component={NewUserPage} />
-        <Route
-          exact
-          path={routes.history}
-          component={() => <div>History</div>}
-        />
 
         <Route exact path="*">
           <Redirect to={routes.dashboard} />
